+++ pdtable/__init__.py
@@ -9,7 +9,7 @@ from .table_metadata import TableMetadata, TableOrigin
 from .store import TableBundle, BlockType, BlockGenerator
 from .units import UnitPolicy
 from .readers.parsers.FixFactory import FixFactory
-from .readers.read_csv import read_csv
+from .readers._csv import read_csv
 from .utils import read_bundle_from_csv, normalized_table_generator
 from .readers.parsers.blocks import JsonData
 from .writers._csv import write_csv
